Moved all settings related files
